Separating month selection view
