+++ kaminotest/kaminotest/Sources/User interface/Main view controller/Graph/BarGraph.swift
